requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
gspread==6.1.0
google-auth==2.35.0
pytz==2024.2
